main Tower
Poop_Tower, Poop_UI, Create_Poop
아군기지 체력, 생성 위치, 초기값

+++ Assets/Script/ScriptableData/Hero_Data.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "HeroData", menuName = "Data")]
 public class Hero_Data : ScriptableObject
 {
-    public int Money;
+    public float Money;
     public UI_Scene.Hero name;
     public Sprite Icon;
     public float Damage;
